Improved load order, new command
Made the admin system initialize for the first time in GM:PostGamemodeLoaded, so that a new hook can be properly utilized; any successive reloads using the reload commands will reload as normal
This new hook allows a server to further fine-tune what is available, and is "BAdmin.AddCommand"; it passes the command name and the settings for it along, and has 3 total returns for controlling it. First is a boolean to overall control whether or not the command should even be added, and the second is a boolean to control overriding settings using the third return
Each of these will modify the console output accordingly, to keep you apprised of any modifications to commands

Also added a new command to use the new(ish) physenv.SetPhysicsPaused function, to overall stop physics from iterating incase someone is being unruly.